+++ workspace/src/localization/localization_shared_utils/setup.cfg
@@ -1,0 +1,4 @@
+[develop]
+script_dir=$base/lib/localization_shared_utils
+[install]
+install_scripts=$base/lib/localization_shared_utils
