+++ workspace/src/localization/particle_filter_estimation/package.xml
@@ -15,7 +15,7 @@
   <exec_depend>python3-matplotlib</exec_depend>
   <exec_depend>python3-numpy</exec_depend>
   <exec_depend version_gte="1.8">python3-scipy</exec_depend>
-  <exec_depend>shared_utils</exec_depend>
+  <exec_depend>localization_shared_utils</exec_depend>
 
   <export>
     <build_type>ament_python</build_type>
