--- workspace/src/localization/shared_utils/setup.cfg
@@ -1,4 +1,0 @@
-[develop]
-script_dir=$base/lib/shared_utils
-[install]
-install_scripts=$base/lib/shared_utils
